requirements.txt
beautifulsoup4==4.6.0
Flask==0.12.2
mechanize==0.3.6
python-dateutil==2.6.1
setuptools==38.2.4
